Migrate WatchVideoChapters to the composition API
